+++ build-binary-from-source.sh
@@ -1,3 +1,3 @@
 cd ..
-R CMD INSTALL --build ggpmisc_0.5.2.tar.gz
+R CMD INSTALL --build ggpmisc_0.5.4-1.9001.tar.gz
 cd ./ggpmisc